requirements.txt
django
pillow
xhtml2pdf
whitenoise
django-cors-headers
djangorestframework
djangorestframework-simplejwt